fix(ui/telescope): Fix extension load order
Telescope doesn't setup extensions with configured options
unless all the extensions are loaded _after_ telescope itself is
setup. Which, in hindsight, makes sense but not so obvious when you're
setting it up. Here we setup the infra (later to be split into a common
module) to enable keeping track of to-be-loaded extensions and loading them
after calling telescope.setup.